requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1;python_version<'3.9'
click==8.2.1
dj-database-url==3.0.0
Django==5.2.1
django-cors-headers==4.7.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.0
uvicorn==0.34.3
whitenoise==6.9.0
